moved evidence estimation to different file in sample
