+++ test/optimize/test_optimize.py
@@ -26,7 +26,6 @@ from pypesto.optimize.ess import (
 )
 from pypesto.optimize.util import (
     assign_ids,
-    laplace_approximation_log_evidence,
 )
 from pypesto.store import read_result
 
@@ -678,25 +677,3 @@ def test_assign_ids():
 
     ids = assign_ids(n_starts=n_starts, ids=None, result=result)
     assert ids == [str(i) for i in range(n_starts, n_starts * 2)]
-
-
-def test_laplace_approximation_log_evidence():
-    """Test the laplace approximation of the log evidence."""
-    log_evidence_true = -1.15  # approximated by hand
-
-    prob = pypesto.Problem(
-        objective=rosen_for_sensi(max_sensi_order=2)["obj"],
-        lb=0 * np.ones((1, 2)),
-        ub=1 * np.ones((1, 2)),
-    )
-
-    # hess
-    result = optimize.minimize(
-        problem=prob,
-        n_starts=10,
-        progress_bar=False,
-    )
-    log_evidence = laplace_approximation_log_evidence(
-        prob, result.optimize_result.x[0]
-    )
-    assert np.isclose(log_evidence, log_evidence_true, atol=0.1)
